waste-dashboard/requirements.txt
# Dashboard Requirements

# Core dependencies
streamlit==1.25.0
pandas==2.0.3
numpy==1.24.3

# Data visualization
altair==5.1.2
plotly==5.18.0
plotly-express==0.4.1
matplotlib==3.8.0
streamlit-plotly-events==0.0.6

# Mapping
folium==0.14.0
streamlit-folium==0.13.0

# Image processing
pillow==10.0.0

# Network and API
requests==2.31.0
paho-mqtt==1.6.1

# Database
pymysql==1.0.3
sqlalchemy==2.0.23
cryptography==41.0.3  # For secure database connections

# Date and time
pytz==2023.3

# Standard library modules are included by default:
# - datetime
# - logging
# - queue
# - json
# - io
# - socket
# - threading
# - atexit
# - os
# - time